+++ src/modes/default.ts
@@ -1,22 +1,8 @@
-import { typingInterval } from '@svelte-typewriter/helpers'
-import type { TypewriterEffectFn, TypewriterModeFn } from '@svelte-typewriter/types'
-
-// Unify the function below with the one on default mode
-const typewriterEffect: TypewriterEffectFn = async ({ currentNode, text }, options) => {
-	currentNode.classList.add('typing')
-	for (let index = 0; index <= text.length; index++) {
-		const char = text[index]
-		char === '<' && (index = text.indexOf('>', index))
-		currentNode.innerHTML = text.slice(0, index)
-		await typingInterval(options.interval)
-	}
-	currentNode.nextSibling !== null && currentNode.classList.length == 1
-		? currentNode.removeAttribute('class')
-		: currentNode.classList.remove('typing')
-}
+import { writeEffect } from '@svelte-typewriter/helpers'
+import type { TypewriterModeFn } from '@svelte-typewriter/types'
 
 const defaultMode: TypewriterModeFn = async ({ elements }, options) => {
-	for (const element of elements) typewriterEffect(element, options)
+	for (const element of elements) writeEffect(element, options)
 
 	// Modularize this
 	const { currentNode: lastElementToFinish } = elements.reduce((longestTextElement, element) => {